pii-redaction-backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
pdfplumber>=0.7.6
python-multipart>=0.0.6
python-dotenv>=1.0.0
# Optional (for OCR of image files)
pillow>=10.0.0
pytesseract>=0.3.10
